feat: Add empty core.management package for custom commands
This commit introduces the `core.management` package. This package serves
as a designated location for storing custom Django management commands.
By creating this package, we establish a structured way to organize and
extend the management capabilities of the core application. Future custom
management commands will be added to this package.