feat: Add WSGI configuration for Django datawarehouse project
This commit adds the `wsgi.py` file, which configures the Web
Server Gateway Interface (WSGI) for the Django `datawarehouse`
project. It sets the `DJANGO_SETTINGS_MODULE` environment variable
to point to the project's settings file and exposes the Django
application as a WSGI callable. This allows the Django application
to be served by WSGI-compliant servers.