feat: Define variables for AWS resources and project configuration
This commit introduces the `variables.tf` file, which defines the input variables required for provisioning AWS resources and configuring the project.

The variables include:

- `aws_region`: AWS region for resource deployment.
- `project_name`: Project name for resource naming.
- `environment`: Deployment environment (dev, staging, production).
- `redshift_db_name`: Redshift database name.
- `redshift_username`: Redshift master username.
- `redshift_password`: Redshift master password (sensitive).
- `redshift_node_type`: Redshift node type.
- `redshift_cluster_type`: Redshift cluster type (single-node or multi-node).
- `redshift_nodes`: Number of Redshift compute nodes.
- `create_glue_resources`: Boolean to control the creation of AWS Glue resources.

These variables allow for customization and configuration of the infrastructure and application deployment. The `sensitive = true` attribute is used for the `redshift_password` variable to prevent it from being displayed in the Terraform state file.